% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emta.andmebaas}
\alias{emta.andmebaas}
\title{EMTA andmebaas}
\format{
A data frame with 2116029 rows and 17 variables:
\describe{
  \item{\code{Kuupäev}}{double Kuupäev}
  \item{\code{aasta}}{integer Aasta}
  \item{\code{kv}}{integer Kvartal}
  \item{\code{Registrikood}}{character Ettevõtte registrikood}
  \item{\code{Nimi}}{character Ettevõtte nimi}
  \item{\code{Liik}}{integer Ettevõtte tüüp}
  \item{\code{KMKR}}{integer Kas on käibemaksukohuslane}
  \item{\code{EMTAK.kood}}{integer EMTAK 1 taseme kood}
  \item{\code{EMTAK}}{integer EMTAK 1 taseme kirjeldus}
  \item{\code{Omavalitsus}}{integer Omavalitsus}
  \item{\code{Maakond}}{integer Maakond}
  \item{\code{Vald}}{integer Vald}
  \item{\code{Linn}}{boolean Kas on linn (jah/ei)}
  \item{\code{Maksud}}{double Tasutud maksud}
  \item{\code{Tööjõumaksud}}{double Tasutud tööjõumaksud}
  \item{\code{Käive}}{double Käive}
  \item{\code{Töötajad}}{double Töötajate arv}
}
}
\usage{
emta.andmebaas
}
\description{
EMTA avalik maksulaekumiste andmebaas
}
\details{
EMTA avalike andmete põhine maksulaekumiste andmebaas
}
\keyword{datasets}
