% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emta.tootajad.kasv}
\alias{emta.tootajad.kasv}
\title{EMTA tööturu andmed}
\format{

}
\usage{
emta.tootajad.kasv
}
\description{
EMTA maksulaekumiste andmed tööturu kohta
}
\details{
DETAILS
}
\keyword{datasets}
