% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ev.emtak}
\alias{ev.emtak}
\title{Ettevõtted EMTAK koodide järgi}
\format{
A data frame with 206314 rows and 9 variables:
}
\source{
https://www.stat.ee/et/esita-andmeid/andmete-esitamisest/ettevotete-uuringud/majandusuksuste-klassifitseerimise-abiinfo
\describe{
  \item{\code{Registrikood}}{character Ettevõtte registrikood}
  \item{\code{Nimi}}{character Ettevõtte nimi}
  \item{\code{EMTAK.1}}{integer EMTAK 1. taseme kood}
  \item{\code{EMTAK.2}}{integer EMTAK 2. taseme kood}
  \item{\code{Klassifikaator}}{integer ESA institutsionaalne sektor. Koodid on toodud andmetabelis "meta.sektor" }
  \item{\code{Emitent}}{integer Emitendi kood. Koodid on toodud andmetabelis "meta.emitent"}
  \item{\code{Registreeritud}}{double Ettevõtte registreerimiskuupäev}
  \item{\code{Kustutatud}}{double Ettevõtte kustutamise kuupäev}
  \item{\code{Muudatus}}{double Ettevõtte andmete muutmise kuupäev}
}
}
\usage{
ev.emtak
}
\description{
Ettevõtete andmed seotud EMTAK koodidega
}
\details{
Andmebaas Ettevõtete EMTAK koodidega
}
\author{
Raoul Lättemäe, jaanuar 2021
}
\keyword{datasets}
