% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ev.register}
\alias{ev.register}
\title{Äriregister}
\format{
A data frame with 311205 rows and 7 variables:
\describe{
  \item{\code{Nimi}}{character Ärinimi}
  \item{\code{Registrikood}}{character Registrikood}
  \item{\code{Staatus}}{integer Staatus koodina}
  \item{\code{Tüüp}}{integer Äriühingu tüüp}
  \item{\code{Staatus.tekstina}}{integer Staatus tekstina}
  \item{\code{Asutamiskuupäev}}{double Asutamiskuupäev}
  \item{\code{Asukoht}}{integer Asukoha kood}
}
}
\usage{
ev.register
}
\description{
Äriregistri andmed (27.01.2021 seisuga)
}
\details{
Äriregistri andmed kõigi äriregistrisse kantud ettevõtete kohta
}
\keyword{datasets}
