% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta.emitent}
\alias{meta.emitent}
\title{Emitendid/Kliendid}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{\code{kood}}{double ev.emtak failis kasutatav emitendi/kliendi kood}
  \item{\code{emitent}}{character Emitendi kirjeldus}
}
}
\source{
https://www.riigiteataja.ee/akt/103062014001
}
\usage{
meta.emitent
}
\description{
Emitentide/Klientide koodid
}
\details{
Emitentide koodide kirjelds
}
\keyword{datasets}
