% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta.emtak}
\alias{meta.emtak}
\title{EMTAK 2008}
\format{
A data frame with 1771 rows and 3 variables:
\describe{
  \item{\code{kood}}{character EMTAK 2008 kood}
  \item{\code{vanem}}{character EMTAK 2008 kood, mille alla konkreetne kood kuulub}
  \item{\code{nimi}}{character Tegevusala kirjeldus}
}
}
\usage{
meta.emtak
}
\description{
EMTAK 2008 koodide kirjeldus
}
\details{
EMTAK 2008 koodid
}
\keyword{datasets}
