% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta.sektor}
\alias{meta.sektor}
\title{Sektorid}
\format{
A data frame with 96 rows and 2 variables:
\describe{
  \item{\code{kood}}{character sektori kood}
  \item{\code{sektor}}{character Sektori nimetus}
}
}
\source{
http://metaweb.stat.ee/view_xml_linear.htm?id=3798567&siteLanguage=ee
}
\usage{
meta.sektor
}
\description{
ESA: institutsionaalsete sektorite (S) klassifikaator 2010.
}
\details{
ESA: institutsionaalsete sektorite (S) klassifikaator 2010.
}
\keyword{datasets}
